REFACTOR: Colocando el componente dentro de la carpeta de seccionees
