+++ src/app/app-routing.routes.ts
@@ -4,9 +4,9 @@ import { AboutComponentRoute } from './components/about/about.component';
 import { ContactComponent } from './components/contact/contact.component';
 import { ExperienceComponent } from './components/experience/experience.component';
 import { HomeComponent } from './components/home/home.component';
-import { PrivacyNoticeComponent } from './components/privacy-notice/privacy-notice.component';
 import { AwardsComponent } from './components/sections/awards/awards.component';
 import { OpticalFiberComponent } from './components/sections/optical-fiber/optical-fiber.component';
+import { PrivacyNoticeComponent } from './components/sections/privacy-notice/privacy-notice.component';
 import { StoriesComponent } from './components/sections/stories/stories.component';
 import { ServicesComponent } from './components/services/services.component';
