+++ src/app/components/sections/privacy-notice/privacy-notice.component.ts
